refactor: Refactored some names for improved readbility
